+++ packages/invalid-of-type-selectors/CHANGELOG.md
@@ -1,16 +1,20 @@
 # Changelog
 
+## 2.0.1
+
+- Fix plugin option `false`
+
 ## 2.0.0
 
-Change rule name to `@mrhenry/stylelint-mrhenry-invalid-of-type-selectors`
+- Change rule name to `@mrhenry/stylelint-mrhenry-invalid-of-type-selectors`
 
 ## 1.0.2
 
-Add support for `stylelint` `15.0.0`
+- Add support for `stylelint` `15.0.0`
 
 ## 1.0.1
 
-Tweak README.md
+- Tweak README.md
 
 ## 1.0.0